Driving from one square to another should be fine?

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Team358/AutoLegacy/Robot358Main.java
@@ -186,7 +186,7 @@ public abstract class Robot358Main extends LinearOpMode {
         }
     }
 
-    public void forward(double power, int distance) {
+    public void forward(double power, double distance) {
 
         //Distance is in inches
